requirements.txt
opencv
pytorch
torchvision
pillow